Fix MDL-82427 deprecation by moving filter.php to classes\text_filter
This change ensures backwardscompatibility to pre-4.5.
